Revert to initializing all weights up front
We can either provide direction with accurate initial weights or
with explore time heuristics. If calculate all the initial weights
O(m*n) up front we get excellent initial direction with the expense
of slower start up and higher memory usage. However, we tend to
solve the problem very quickly and accurately once we have that.
Lazy weight initialization is efficient only if the scorer is applied
then.